11.5.0，群聊界面、逻辑完善

+++ App/qintalker/factory/src/main/java/net/lcadsl/qintalker/factory/presenter/message/ChatContract.java
@@ -40,6 +40,6 @@ public interface ChatContract {
         //显示管理员菜单
         void showAdminOption(boolean isAdmin);
         //初始化成员
-        void onInitGroupMembers(List<MemberUserModel> members,int moreCount);
+        void onInitGroupMembers(List<MemberUserModel> members,long moreCount);
     }
 }
